+++ Projects/CursoPZ/Class_4/main.cpp
@@ -204,8 +204,8 @@ int main() {
 }
 
 TPZGeoMesh *GetMesh (int nx,int ny) {
-	int i,j;
-	int id, index;
+	long i,j;
+	long id, index;
 	
 	//Let's try with an unitary domain
 	REAL lx = 1.;
@@ -235,7 +235,7 @@ TPZGeoMesh *GetMesh (int nx,int ny) {
 	}
 
 	//Auxiliar vector to store a element connectivities
-	TPZVec <int> connect(4,0);
+	TPZVec <long> connect(4,0);
 	
 	//Element connectivities
 	for(i = 0; i < (nx - 1); i++){
